Refactored Migrations and seeder
