--- database/seeders/BlogPostSeeder.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Database\Seeders;
-
-use App\Models\BlogPost;
-use Illuminate\Database\Seeder;
-
-class BlogPostSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        BlogPost::factory()->count(10)->create();
-    }
-}
