--- database/seeders/MembershipSeeder.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Database\Seeders;
-
-use App\Models\Membership;
-use Illuminate\Database\Seeder;
-
-class MembershipSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        Membership::factory()->count(10)->create();
-    }
-}
