--- database/seeders/NewsCategorySeeder.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Database\Seeders;
-
-use App\Models\NewsCategory;
-use Illuminate\Database\Seeder;
-
-class NewsCategorySeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        NewsCategory::factory()->count(10)->create();
-
-    }
-}
