--- database/seeders/NewsPostSeeder.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Database\Seeders;
-
-use App\Models\NewsPost;
-use Illuminate\Database\Seeder;
-
-class NewsPostSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        NewsPost::factory()->count(50)->create();
-
-    }
-}
